pyproject.toml
[project]
name = "fastapiproject1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.66.0",
    "deepgram-sdk==2.12.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
